backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers==4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "faker>=40.5.1",
    "mistralai>=1.12.4",
]
